refactor(ivy): rename parser folder and files
