+++ packages/compiler-cli/src/ngcc/src/parsing/file_parser.ts
