+++ packages/compiler-cli/src/ngcc/src/parsing/parsed_class.ts
